Fix Projects Tests and Contract:
- [Contract] Replace bytes20 and bytes12 by bytes32 for owner and repo*
- [Test] Modify the addRepo test case to check the real Github values
- [Test] Initialize the app and modify imports and initialization hooks to resemble allocations test file

*The owner and repo fields don't fit encoded in less than bytes32 since we need to decode, so truncating the value is not a good idea
